Move the event channel to the base activity
