+++ src/android/app/src/main/kotlin/com/nivisi/leafy_launcher/NotesActivity.kt
@@ -40,6 +40,6 @@ class NotesActivity: LeafyActivityBase() {
     }
 
     companion object {
-        lateinit var self: NotesActivity
+        var self: NotesActivity? = null
     }
 }
