+++ src/android/app/src/main/kotlin/com/nivisi/leafy_launcher/broadcast_receivers/DeviceLocaleChangedBroadcastReceiver.kt
@@ -4,15 +4,15 @@ import android.content.BroadcastReceiver
 import android.content.Context
 import android.content.Intent
 import com.nivisi.leafy_launcher.MainActivity
+import com.nivisi.leafy_launcher.NotesActivity
 import java.util.*
 
 class DeviceLocaleChangedBroadcastReceiver : BroadcastReceiver() {
-
     override fun onReceive(context: Context?, intent: Intent?) {
         if (intent?.action == Intent.ACTION_LOCALE_CHANGED) {
-            MainActivity.self.dispatchDeviceLocaleChangedEvent(
-                Locale.getDefault().toLanguageTag()
-            )
+            val languageTag = Locale.getDefault().toLanguageTag()
+            MainActivity.self?.dispatchDeviceLocaleChangedEvent(languageTag)
+            NotesActivity.self?.dispatchDeviceLocaleChangedEvent(languageTag)
         }
     }
 }
